+++ lib/src/cook/chai/Toolchain.cpp
@@ -60,10 +60,18 @@ namespace cook { namespace chai {
         {
             return manager_->has_config(key);
         }
-
         bool Toolchain::has_config_2(const std::string & key, const std::string & value) const
         {
             return manager_->has_config(key, value);
+        }
+
+        bool Toolchain::remove_config_1(const std::string & key)
+        {
+            return manager_->remove_config(key);
+        }
+        bool Toolchain::remove_config_2(const std::string & key, const std::string & value)
+        {
+            return manager_->remove_config(key, value);
         }
 
         void Toolchain::configure(unsigned int priority, const std::string & uuid, const ConfigurationCallback & cb)