+++ lib/src/cook/process/toolchain/Types.hpp
@@ -14,7 +14,7 @@ namespace cook { namespace process { namespace toolchain {
     enum class Part
     {
         Begin_,
-        Cli = Begin_, Pre, Deps, Export, Output, Input, DepFile, Option, Define, IncludePath, ForceInclude, Library, LibraryPath, Framework, FrameworkPath, Resource,
+        Cli = Begin_, Pre, Runtime, Deps, Export, Output, Input, DepFile, Option, Define, IncludePath, ForceInclude, Library, LibraryPath, Framework, FrameworkPath, Resource,
         End_
     };
 
@@ -25,6 +25,7 @@ namespace cook { namespace process { namespace toolchain {
 #define L_CASE(NAME) case Part::NAME: return oss << #NAME
             L_CASE(Cli);
             L_CASE(Pre);
+            L_CASE(Runtime);
             L_CASE(Deps);
             L_CASE(Export);
             L_CASE(Output);