+++ lib/src/cook/process/toolchain/serialize/MSVC.cpp
@@ -16,6 +16,7 @@ namespace cook { namespace process { namespace toolchain { namespace serialize {
         if (false) {
         } else if (k == "debug_symbols" && v == "true") {
             kv.append(Part.Pre, "Zi")
+            tm[Part.Runtime] = fun(k,v) { return "/${k}d" }
         } else if (k == "optimization" && v == "max_speed") {
             kv.append(Part.Pre, "O2")
         } else if (k == "arch" && v == "x86") {
@@ -23,8 +24,8 @@ namespace cook { namespace process { namespace toolchain { namespace serialize {
         } else if (k == "position_independent_code" && v == "true") {
         } else if (k == "c++.runtime") {
             if (false) {}
-            else if (v == "dynamic") { kv.append(Part.Pre, "MD") }
-            else if (v == "static") { kv.append(Part.Pre, "MT") }
+            else if (v == "dynamic") { kv.append(Part.Runtime, "MD") }
+            else if (v == "static") { kv.append(Part.Runtime, "MT") }
         } else if (k == "c++.std" && e.language == Language.CXX) {
             kv.append(Part.Pre, "std", "c++${v}")
         } else {
@@ -55,6 +56,7 @@ for( lang : [Language.C, Language.CXX, Language.ASM, Language.Resource]) {
 
     tm[Part.Cli]            = fun(k,v) { return k }
     tm[Part.Pre]            = fun(k,v) { if (v.empty) { return "/${k}" } else { return "/${k}:${v}" } }
+    tm[Part.Runtime]        = fun(k,v) { return "/${k}" }
     tm[Part.Output]         = fun(k,v) { return "/Fo${k}" }
     tm[Part.Input]          = fun(k,v) { return k }
     tm[Part.Deps]           = fun(k,v) { return "/showIncludes" }
@@ -69,6 +71,7 @@ for( lang : [Language.C, Language.CXX, Language.ASM, Language.Resource]) {
     if (lang == Language.Resource) {
         kv.append(Part.Cli, "rc")
         tm[Part.Pre]        = fun(k,v) { return "" }
+        tm[Part.Runtime]    = fun(k,v) { return "" }
         tm[Part.Deps]       = fun(k,v) { return "" }
         tm[Part.Resource]   = fun(k,v) { return k }
     } else {